backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-ollama==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
unstructured==0.10.30

